src/app/agent-call-stats/call-stats-agent-view/call-stats-agent-view.component.spec.ts

import { ComponentFixture, TestBed } from '@angular/core/testing';
import {MatSelectModule} from '@angular/material/select';
import {MatPaginatorModule} from '@angular/material/paginator';
import {MatTableModule} from '@angular/material/table';
import {HttpClientModule} from '@angular/common/http';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import {MatDialogModule} from '@angular/material/dialog';
import { MatDatepickerModule } from '@angular/material/datepicker';
import { MatNativeDateModule } from '@angular/material/core';
import { MatInputModule } from '@angular/material/input';
import { MatFormFieldModule } from "@angular/material/form-field";
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { CallStatsAgentViewComponent } from './call-stats-agent-view.component';

describe('CallStatsAgentViewComponent', () => {
  let component: CallStatsAgentViewComponent;
  let fixture: ComponentFixture<CallStatsAgentViewComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ CallStatsAgentViewComponent ],
      imports:[
        MatDatepickerModule,
        MatFormFieldModule,
        MatSelectModule,
        MatPaginatorModule,
        MatTableModule,
        HttpClientModule,
        FormsModule, 
        ReactiveFormsModule,
        MatDialogModule,
        MatNativeDateModule,
        MatInputModule,
        HttpClientModule,
        BrowserAnimationsModule
      ]
    })
    .compileComponents();

    fixture = TestBed.createComponent(CallStatsAgentViewComponent);
  });

  it('should create', () => {
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  });

  it('should have type drop down', ()=>{
    fixture.detectChanges();
    const app = fixture.debugElement.nativeElement as HTMLElement;
    expect(app.querySelector('#typeDropDown')).toBeTruthy();
  })

  it('should have summary table', ()=>{
    fixture.detectChanges();
    const app = fixture.debugElement.nativeElement as HTMLElement;
    expect(app.querySelector('#summaryTable')).toBeTruthy();
  })

  it('should have call logs table', ()=>{
    fixture.detectChanges();
    const app = fixture.debugElement.nativeElement as HTMLElement;
    expect(app.querySelector('#calllogsTable')).toBeTruthy();
  })
});
